deployment/requirements.txt
firebase-admin
python-dotenv
streamlit
